+++ src/anemoi/transform/filters/icon_refinement_level.py
@@ -9,15 +9,17 @@
 
 
 import logging
+from typing import Any
 
+import earthkit.data as ekd
 import tqdm
 
-from ..fields import new_field_from_latitudes_longitudes
-from ..fields import new_field_from_numpy
-from ..fields import new_fieldlist_from_list
-from ..filter import Filter
-from ..grids.icon import icon_grid
-from . import filter_registry
+from anemoi.transform.fields import new_field_from_latitudes_longitudes
+from anemoi.transform.fields import new_field_from_numpy
+from anemoi.transform.fields import new_fieldlist_from_list
+from anemoi.transform.filter import Filter
+from anemoi.transform.filters import filter_registry
+from anemoi.transform.grids.icon import icon_grid
 
 LOG = logging.getLogger(__name__)
 
@@ -26,14 +28,36 @@ LOG = logging.getLogger(__name__)
 class IconRefinement(Filter):
     """A filter interpolate its input to an ICON grid."""
 
-    def __init__(self, *, grid, refinement_level_c):
+    def __init__(self, *, grid: Any, refinement_level_c: Any) -> None:
+        """Initialize the IconRefinement filter.
+
+        Parameters
+        ----------
+        grid : Any
+            The grid to use for interpolation.
+        refinement_level_c : Any
+            The refinement level for the grid.
+        """
+
         self.grid = grid
         self.refinement_level_c = refinement_level_c
 
         self.latitudes, self.longitudes = icon_grid(self.grid, self.refinement_level_c)
         self.nearest_grid_points = None
 
-    def forward(self, fields):
+    def forward(self, fields: ekd.FieldList) -> ekd.FieldList:
+        """Interpolate the input fields to an ICON grid.
+
+        Parameters
+        ----------
+        fields : ekd.FieldList
+            List of fields to be interpolated.
+
+        Returns
+        -------
+        ekd.FieldList
+            List of interpolated fields.
+        """
         if self.nearest_grid_points is None:
             from anemoi.utils.grids import nearest_grid_points
 
@@ -58,6 +82,3 @@ class IconRefinement(Filter):
             )
 
         return new_fieldlist_from_list(result)
-
-    def backward(self, data):
-        raise NotImplementedError("IconRefinement is not reversible")