+++ src/anemoi/transform/filters/land_parameters.py
@@ -15,8 +15,9 @@ from typing import List
 import earthkit.data as ekd
 import numpy as np
 
-from . import filter_registry
-from .base import SimpleFilter
+from anemoi.transform.filters import filter_registry
+from anemoi.transform.filters.matching import MatchingFieldsFilter
+from anemoi.transform.filters.matching import matching
 
 SOIL_TYPE_DIC = {
     0: {"theta_pwp": 0, "theta_cap": 0},
@@ -61,12 +62,12 @@ def read_crosswalking_table(param: Any, param_dic: Dict[int, Dict[str, float]]) 
     ----------
     param : Any
         The parameter to read.
-    param_dic : dict
+    param_dic : Dict[int, Dict[str, float]]
         The dictionary containing the crosswalking table.
 
     Returns
     -------
-    list of np.ndarray
+    List[np.ndarray]
         The arrays for each key in the crosswalking table.
     """
     arrays = [np.array([param_dic[x][key] for x in param]) for key in param_dic[0].keys()]
@@ -74,9 +75,13 @@ def read_crosswalking_table(param: Any, param_dic: Dict[int, Dict[str, float]]) 
 
 
 @filter_registry.register("land_parameters")
-class LandParameters(SimpleFilter):
+class LandParameters(MatchingFieldsFilter):
     """A filter to add static parameters from table based on soil/vegetation type."""
 
+    @matching(
+        select="param",
+        forward=("high_veg_type", "low_veg_type", "soil_type"),
+    )
     def __init__(
         self,
         *,
@@ -92,6 +97,7 @@ class LandParameters(SimpleFilter):
         theta_pwp: str = "theta_pwp",
         theta_cap: str = "theta_cap",
     ) -> None:
+
         self.high_veg_type = high_veg_type
         self.low_veg_type = low_veg_type
         self.soil_type = soil_type
@@ -104,37 +110,21 @@ class LandParameters(SimpleFilter):
         self.theta_pwp = theta_pwp
         self.theta_cap = theta_cap
 
-    def forward(self, data: ekd.FieldList) -> ekd.FieldList:
-        """Forward transformation to add static parameters.
-
-        Parameters
-        ----------
-        data : Any
-            The input data.
-
-        Returns
-        -------
-        Any
-            The transformed data.
-        """
-        return self._transform(
-            data,
-            self.forward_transform,
-            self.high_veg_type,
-            self.low_veg_type,
-            self.soil_type,
-        )
-
-    def forward_transform(self, tvh: ekd.Field, tvl: ekd.Field, sotype: ekd.Field) -> Iterator[ekd.Field]:
+    def forward_transform(
+        self,
+        high_veg_type: ekd.Field,
+        low_veg_type: ekd.Field,
+        soil_type: ekd.Field,
+    ) -> Iterator[ekd.Field]:
         """Get static parameters from table based on soil/vegetation type.
 
         Parameters
         ----------
-        tvh : ekd.Field
+        high_veg_type : ekd.Field
             High vegetation type.
-        tvl : ekd.Field
+        low_veg_type : ekd.Field
             Low vegetation type.
-        sotype : ekd.Field
+        soil_type : ekd.Field
             Soil type.
 
         Returns
@@ -142,15 +132,15 @@ class LandParameters(SimpleFilter):
         Iterator[ekd.Field]
             An iterator over the new fields with static parameters.
         """
-        hveg_rsmin, hveg_cov, hveg_z0m = read_crosswalking_table(tvh.to_numpy(), VEG_TYPE_DIC)
-        lveg_rsmin, lveg_cov, lveg_z0m = read_crosswalking_table(tvl.to_numpy(), VEG_TYPE_DIC)
-        theta_pwp, theta_cap = read_crosswalking_table(sotype.to_numpy(), SOIL_TYPE_DIC)
+        hveg_rsmin, hveg_cov, hveg_z0m = read_crosswalking_table(high_veg_type.to_numpy(), VEG_TYPE_DIC)
+        lveg_rsmin, lveg_cov, lveg_z0m = read_crosswalking_table(low_veg_type.to_numpy(), VEG_TYPE_DIC)
+        theta_pwp, theta_cap = read_crosswalking_table(soil_type.to_numpy(), SOIL_TYPE_DIC)
 
-        yield self.new_field_from_numpy(hveg_rsmin, template=tvh, param=self.hveg_rsmin)
-        yield self.new_field_from_numpy(hveg_cov, template=tvh, param=self.hveg_cov)
-        yield self.new_field_from_numpy(hveg_z0m, template=tvh, param=self.hveg_z0m)
-        yield self.new_field_from_numpy(lveg_rsmin, template=tvl, param=self.lveg_rsmin)
-        yield self.new_field_from_numpy(lveg_cov, template=tvl, param=self.lveg_cov)
-        yield self.new_field_from_numpy(lveg_z0m, template=tvl, param=self.lveg_z0m)
-        yield self.new_field_from_numpy(theta_pwp, template=sotype, param=self.theta_pwp)
-        yield self.new_field_from_numpy(theta_cap, template=sotype, param=self.theta_cap)
+        yield self.new_field_from_numpy(hveg_rsmin, template=high_veg_type, param=self.hveg_rsmin)
+        yield self.new_field_from_numpy(hveg_cov, template=high_veg_type, param=self.hveg_cov)
+        yield self.new_field_from_numpy(hveg_z0m, template=high_veg_type, param=self.hveg_z0m)
+        yield self.new_field_from_numpy(lveg_rsmin, template=low_veg_type, param=self.lveg_rsmin)
+        yield self.new_field_from_numpy(lveg_cov, template=low_veg_type, param=self.lveg_cov)
+        yield self.new_field_from_numpy(lveg_z0m, template=low_veg_type, param=self.lveg_z0m)
+        yield self.new_field_from_numpy(theta_pwp, template=soil_type, param=self.theta_pwp)
+        yield self.new_field_from_numpy(theta_cap, template=soil_type, param=self.theta_cap)