+++ src/anemoi/transform/filters/snow_cover.py
@@ -7,14 +7,14 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-from typing import Any
 from typing import Iterator
 
 import earthkit.data as ekd
 import numpy as np
 
-from . import filter_registry
-from .base import SimpleFilter
+from anemoi.transform.filters import filter_registry
+from anemoi.transform.filters.matching import MatchingFieldsFilter
+from anemoi.transform.filters.matching import matching
 
 
 def compute_snow_cover(snow_depth: np.ndarray, snow_density: np.ndarray) -> np.ndarray:
@@ -40,9 +40,13 @@ def compute_snow_cover(snow_depth: np.ndarray, snow_density: np.ndarray) -> np.n
 
 
 @filter_registry.register("snow_cover")
-class SnowCover(SimpleFilter):
+class SnowCover(MatchingFieldsFilter):
     """A filter to compute snow cover from snow density and snow depth."""
 
+    @matching(
+        select="param",
+        forward=("snow_depth", "snow_density"),
+    )
     def __init__(
         self,
         *,
@@ -61,38 +65,19 @@ class SnowCover(SimpleFilter):
         snow_cover : str, optional
             The parameter name for snow cover, by default "snowc".
         """
+
         self.snow_depth = snow_depth
         self.snow_density = snow_density
         self.snow_cover = snow_cover
 
-    def forward(self, data: ekd.FieldList) -> ekd.FieldList:
-        """Apply the forward transformation to the data.
-
-        Parameters
-        ----------
-        data : Any
-            The input data.
-
-        Returns
-        -------
-        Any
-            The transformed data.
-        """
-        return self._transform(
-            data,
-            self.forward_transform,
-            self.snow_depth,
-            self.snow_density,
-        )
-
-    def forward_transform(self, sd: Any, rsn: Any) -> Iterator[ekd.Field]:
+    def forward_transform(self, snow_depth: ekd.Field, snow_density: ekd.Field) -> Iterator[ekd.Field]:
         """Convert snow depth and snow density to snow cover.
 
         Parameters
         ----------
-        sd : Any
+        snow_depth : ekd.Field
             The snow depth data.
-        rsn : Any
+        snow_density : ekd.Field
             The snow density data.
 
         Returns
@@ -100,6 +85,6 @@ class SnowCover(SimpleFilter):
         Iterator[ekd.Field]
             Transformed fields.
         """
-        snow_cover = compute_snow_cover(sd.to_numpy(), rsn.to_numpy())
+        snow_cover = compute_snow_cover(snow_depth.to_numpy(), snow_density.to_numpy())
 
-        yield self.new_field_from_numpy(snow_cover, template=sd, param=self.snow_cover)
+        yield self.new_field_from_numpy(snow_cover, template=snow_depth, param=self.snow_cover)