+++ CHANGELOG.md
@@ -8,6 +8,58 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please add your functional changes to the appropriate section in the PR.
 Keep it human-readable, your future self will thank you!
 
+## [0.1.5](https://github.com/ecmwf/anemoi-transform/compare/0.1.4...0.1.5) (2025-03-28)
+
+
+### Bug Fixes
+
+* missing metadata attribute ([#68](https://github.com/ecmwf/anemoi-transform/issues/68)) ([4f69cb4](https://github.com/ecmwf/anemoi-transform/commit/4f69cb480cc09c1b9b466a81b671b8427c87866d))
+
+
+### Documentation
+
+* Docathon ([#62](https://github.com/ecmwf/anemoi-transform/issues/62)) ([413665c](https://github.com/ecmwf/anemoi-transform/commit/413665cf8b475bbf673017bca66b9b1360ded4ea))
+
+## [0.1.4](https://github.com/ecmwf/anemoi-transform/compare/0.1.3...0.1.4) (2025-03-24)
+
+
+### Features
+
+* plugin support ([#65](https://github.com/ecmwf/anemoi-transform/issues/65)) ([7481145](https://github.com/ecmwf/anemoi-transform/commit/7481145c51f4fbf2fdd43a9b8822b18e32b62449))
+* Rodeo transform update ([#61](https://github.com/ecmwf/anemoi-transform/issues/61)) ([3e669c0](https://github.com/ecmwf/anemoi-transform/commit/3e669c0207c68b897126e128a76d82a921960522))
+
+
+### Bug Fixes
+
+* fix regrid arguments ([#67](https://github.com/ecmwf/anemoi-transform/issues/67)) ([1a730cd](https://github.com/ecmwf/anemoi-transform/commit/1a730cd6354cdd00eedc82ec5bf57eea34e8f797))
+* return unmodified fields ([#57](https://github.com/ecmwf/anemoi-transform/issues/57)) ([6f0e6f4](https://github.com/ecmwf/anemoi-transform/commit/6f0e6f46506f8eef28219a26e1cfddca6b81793d))
+* undetected as dry pixels and meaningful naming ([#59](https://github.com/ecmwf/anemoi-transform/issues/59)) ([b66ce25](https://github.com/ecmwf/anemoi-transform/commit/b66ce25af3f1affc5c8162567a5754dc91b14889))
+
+
+### Documentation
+
+* links to GitHub ([#66](https://github.com/ecmwf/anemoi-transform/issues/66)) ([10f5c13](https://github.com/ecmwf/anemoi-transform/commit/10f5c138a6c96f1420d476d0171c5f1850cac43b))
+
+## [0.1.3](https://github.com/ecmwf/anemoi-transform/compare/0.1.2...0.1.3) (2025-03-06)
+
+
+### Features
+
+* Add `icon_refinement_level` filter ([#32](https://github.com/ecmwf/anemoi-transform/issues/32)) ([a2c0711](https://github.com/ecmwf/anemoi-transform/commit/a2c07114c18e6b631401f92519bd760564f8e1ac))
+* Clipping strategy to solve issues in the ORAS6 reanalysis ([#56](https://github.com/ecmwf/anemoi-transform/issues/56)) ([179d602](https://github.com/ecmwf/anemoi-transform/commit/179d602f84ad87d5f7d6fe3b7d2348ed74d55a13))
+* Matching fields ([#55](https://github.com/ecmwf/anemoi-transform/issues/55)) ([b6f52d6](https://github.com/ecmwf/anemoi-transform/commit/b6f52d6073b3b5de7e44e5e4694b8eeab0b339c2))
+
+
+### Bug Fixes
+
+* for opera mask ([#40](https://github.com/ecmwf/anemoi-transform/issues/40)) ([a1f8bf8](https://github.com/ecmwf/anemoi-transform/commit/a1f8bf8b49db74d7b79dea0a800d00dabdaa1ba2))
+
+
+### Documentation
+
+* fix readthedocs ([#51](https://github.com/ecmwf/anemoi-transform/issues/51)) ([42e874b](https://github.com/ecmwf/anemoi-transform/commit/42e874b1020f6d542d6bf9d20ee3c43483c2abcb))
+* use new logo ([#42](https://github.com/ecmwf/anemoi-transform/issues/42)) ([aa13ac5](https://github.com/ecmwf/anemoi-transform/commit/aa13ac5b9424d40f5d6bce8279ecbe73292bcc0b))
+
 ## 0.1.2 (2025-02-04)
 
 <!-- Release notes generated using configuration in .github/release.yml at main -->